🌊 Feat: 처음 화면 열 때 댓글 데이터도 받아오게 구현
이미지 여러장 올리면 다른 사진 보이게 수정

추가 변경 사항
refactor : CommentViewModel -> FeedViewModel에 합병
