+++ PADO/PADO/iOS/ContentView.swift
@@ -16,7 +16,6 @@ struct ContentView: View {
     
     @StateObject var surfingVM = SurfingViewModel()
     @StateObject var feedVM = FeedViewModel()
-    @StateObject var commentVM = CommentViewModel()
     @StateObject var followVM = FollowViewModel()
     
     @State private var selectedTab = 0
@@ -33,7 +32,7 @@ struct ContentView: View {
         NavigationStack {
             GeometryReader { geometry in
                 TabView(selection: $selectedTab) {
-                    FeedView(feedVM: feedVM, commentVM: commentVM)
+                    FeedView(feedVM: feedVM)
                         .tabItem {
                             Image(selectedTab == 0 ? "home_light" : "home_gray")
