+++ PADO/PADO/iOS/MainView/Feed/Cell/HeartCommentCell.swift
@@ -15,7 +15,6 @@ struct HeartCommentCell: View {
     @Binding var isShowingCommentView: Bool
     
     @ObservedObject var feedVM: FeedViewModel
-    @ObservedObject var commentVM: CommentViewModel
     
     var body: some View {
         VStack(spacing: 16) {
@@ -45,12 +44,12 @@ struct HeartCommentCell: View {
                     Image("chat")
                 }
                 .sheet(isPresented: $isShowingCommentView) {
-                    CommentView(commentVM: commentVM, feedVM: feedVM)
+                    CommentView(feedVM: feedVM)
                         .presentationDetents([.fraction(0.99), .fraction(0.8)])
                         .presentationDragIndicator(.visible)
                 }
                 // 댓글이 달릴 때 마다 +1 카운팅 되게 하는 로직 추가
-                Text(String(commentVM.comments.count))
+                Text(String(feedVM.comments.count))
                     .font(.system(size: 12))
                     .fontWeight(.semibold)
                     .shadow(radius: 1, y: 1)
