+++ PADO/PADO/iOS/MainView/Feed/View/FeedView.swift
@@ -15,7 +15,6 @@ struct FeedView: View {
     @EnvironmentObject var authenticationViewModel: AuthenticationViewModel
     
     @StateObject var feedVM: FeedViewModel
-    @ObservedObject var commentVM: CommentViewModel
     @StateObject private var mainCommentVM = MainCommentViewModel()
     @StateObject private var mainFaceMojiVM = MainFaceMojiViewModel()
     
@@ -85,7 +84,7 @@ struct FeedView: View {
                     Spacer()
                     
                     // MARK: - HeartComment
-                    HeartCommentCell(isShowingReportView: $feedVM.isShowingReportView, isShowingCommentView: $feedVM.isShowingCommentView, feedVM: feedVM, commentVM: commentVM)
+                    HeartCommentCell(isShowingReportView: $feedVM.isShowingReportView, isShowingCommentView: $feedVM.isShowingCommentView, feedVM: feedVM)
                         .padding(.leading, UIScreen.main.bounds.width)
                         .padding(.trailing, 60)
                         .padding(.bottom, 10)
@@ -95,7 +94,7 @@ struct FeedView: View {
                         ScrollView(.horizontal, showsIndicators: false) {
                             HStack(spacing: 16) {
                                 ForEach(Array(feedVM.stories.enumerated()), id: \.element) { index, story in
-                                    StoryCell(story: story, storyIndex: index, feedVM: feedVM, commentVM: commentVM) {
+                                    StoryCell(story: story, storyIndex: index, feedVM: feedVM) {
                                         self.feedVM.selectStory(story)
                                     }
                                 }
